+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `getQuarter` utility function
-- `getPreviousQuarter` utility function
+- `addQuarters` utility function
 
 ## [2.3.0] - 2026-04-07